+++ packages/snap-networks-utils/src/utils/errors/index.ts
@@ -1,5 +1,14 @@
+export { createSnapErrorHandling, createTrackError } from './trackError';
 export { createWithCatchAndThrowSnapError, normalizeError } from './errors';
 export { isSnapRpcError } from './snapRpcError';
+export type {
+  CreateSnapErrorHandlingOptions,
+  CreateTrackErrorOptions,
+  PrepareErrorForTrackingFn,
+  ShouldTrackErrorFn,
+  SnapTrackErrorRequest,
+  TrackErrorCapableProvider,
+} from './trackError';
 export type {
   CreateWithCatchAndThrowSnapErrorOptions,
   LogErrorFn,